fix: user and post services

+++ app/Http/Services/PostService.php
@@ -7,36 +7,51 @@ use App\DTO\PostDTO;
 use App\Exceptions\BusinessException;
 use App\Exceptions\ModelDeletionException;
 use App\Exceptions\ModelNotFoundException;
+use App\Exceptions\ModelUpdationException;
 use App\Http\Resources\CommentResource;
 use App\Models\Post;
 use App\Repositories\PostRepository;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
 
 class PostService
 {
-    private IPostRepository $repository;
+    /**
+     * @var IPostRepository|PostRepository
+     */
+    private IPostRepository|PostRepository $repository;
 
+    /**
+     *
+     */
     public function __construct()
     {
         $this->repository = new PostRepository();
     }
 
+    /**
+     * @param PostDTO $postDTO
+     * @return Post
+     */
     public function create(PostDTO $postDTO): Post
     {
         return $this->repository->createPost($postDTO);
     }
 
-    public function update(PostDTO $postDTO, int $id): Post
-    {
-        $postWithId = $this->repository->getPostById($id);
-
-        return $this->repository->updatePost($postDTO, $postWithId);
-    }
-
+    /**
+     * @param int $id
+     * @return Post
+     * @throws ModelNotFoundException
+     */
     public function show(int $id): Post
     {
         $postWithId = $this->repository->getPostById($id);
+
+        if (Auth::user()->id !== $postWithId->user_id) {
+            $this->repository->incrementPostViews($postWithId);
+        }
 
         if ($postWithId === null) {
             throw new ModelNotFoundException(__('messages.post_not_found'));
@@ -45,6 +60,39 @@ class PostService
         return $postWithId;
     }
 
+    /**
+     * @param PostDTO $postDTO
+     * @param int $id
+     * @return Post
+     * @throws BusinessException
+     * @throws ModelNotFoundException
+     * @throws ModelUpdationException
+     */
+    public function update(PostDTO $postDTO, int $id): Post
+    {
+        $postWithId = $this->repository->getPostById($id);
+
+        if ($postWithId === null) {
+            throw new ModelNotFoundException(__('messages.post_not_found'));
+        }
+
+        if (Auth::user()->id !== $postWithId->user_id) {
+            throw new BusinessException(
+                __('messages.unauthorized_post_edit'),
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
+        $data = $this->repository->updatePost($postDTO, $postWithId);
+        throw new ModelUpdationException(__('messages.post_updated'), 0, $data);
+    }
+
+    /**
+     * @param int $id
+     * @return Post|JsonResponse
+     * @throws ModelDeletionException
+     * @throws ModelNotFoundException|BusinessException
+     */
     public function delete(int $id): Post|JsonResponse
     {
         $postWithId = $this->repository->getPostById($id);
@@ -53,10 +101,22 @@ class PostService
             throw new ModelNotFoundException(__('messages.record_not_found'));
         }
 
+        if (Auth::user()->id !== $postWithId->user_id) {
+            throw new BusinessException(
+                __('messages.unauthorized_post_delete'),
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
         $this->repository->deletePost($postWithId);
         throw new ModelDeletionException(__('messages.record_deleted'));
     }
 
+    /**
+     * @param int $post_id
+     * @return JsonResponse|AnonymousResourceCollection
+     * @throws ModelNotFoundException
+     */
     public function getComments(
         int $post_id
     ): JsonResponse|AnonymousResourceCollection
@@ -73,6 +133,12 @@ class PostService
         return CommentResource::collection($comments);
     }
 
+    /**
+     * @param int $post_id
+     * @param int $comment_id
+     * @return JsonResponse|CommentResource
+     * @throws ModelNotFoundException
+     */
     public function getCommentById(
         int $post_id,
         int $comment_id,
